Fix flamegraph name matching.
Previously, we were using the same regexp to match both a node's
FunctionName (for server-side filtering) and UniqueName (for
browser-side pivoting). However the two strings can differ (UniqueName
can end with a #<number> in case of overloads). This resulted in us
being unable to ignore/focus/hide overloaded functions.

We now use distinct regexps for the two different use cases.  And we
no longer compute the regexps in Go. Instead we do this in Javascript
using the pre-existing quotemeta function.

The quotemeta function has now been moved to the top-level and renamed
to pprofQuoteMeta.