Use html/template to produce source listing. (#831)
We previously used an ad-hoc combination of calls to printf
and HTML escaping routines to produce HTML for source code
listing. We now use html/template to do the formatting. This
should give us HTML generation that is less susceptible to
escaping issues.

The change keeps the output very similar to the old output
(mostly a small number of spacing changes that do not affect
what a user can see).

Also added a test that does a quick check of the source view
displayed in a browser.