Flamegraph uses seconds instead of hours when appropriate.
The Javascript code now behave the same as the Go code. (It switches
to the next unit only if one unit is at least as big as the value
being displayed.) The main effect is that we no longer use "hours" as
the unit for time values in the range [36s,3600s). We only switch to
using "hours" when the measured time is at least one hour. This is
easier for users to understand: e.g., a time value of hundred seconds
now displays as "100s", not "0.03hrs".

Also added a unittest for the Javascript unit picking code.