Update kernel base to work for non-KASLR kernels. (#904)
Replace an existing check for "start >= loadSegment.Vaddr" with
"start >= 0x8000000000000000".

The old condition was checking for the mapping start address to be greater
or equal to the program segment Vaddr. However, for kernels without KASLR,
the mapping start address (start) is less or equal the program header Vaddr.
If the program header that includes the '.text' section is not the first
segment, then the mapping start is strictly less than loadSegment.Vaddr.

Instead of removing the condition, we replace it with a check that start is
in the kernel address space, so we don't take this path for user space
binaries.

Adjusted the test case for PIE kernels to match actual addresses seen for
a non-KASLR kernel. The new test fails with the old code and succeeds now.