Allow symbolz to try symbolize unsymbolizable mappings. (#907)
We added skipping unsymbolizable mappings in symbolz long time ago in
PR #368 to address #339 where we saw error like "unexpected negative
adjusted address".

But that error was fixed in a more proper way in subsequent #397 to fix
another reported issue #280 (and internal b/32020573). So skipping
unsymbolized mappings shouldn't be needed anymore.

I tested this by verifying that the test case from #339 still works fine
with the proposed change. And that it does fail if I roll back #397
locally.

This change is useful as we experiment with using symbolz to symbolize
JIT locations from //anon (which is unsymbolizable per current terminology).

Co-authored-by: Maggie Nolan Edmonds <nolanmar@google.com>