Fix the OSX demangling fallback comparison.
This is a follow-up to #924. In the comparison we should compare to the
string with stripped underscore, otherwise the comparison will never be
true and we strip the underscore in cases we shouldn't. See the added
test case.

Also, while we are here, add copying demangling options that we also
missed. Hopefully we can get
https://github.com/ianlancetaylor/demangle/pull/24 in and this won't be
needed.