Make flame graph the default view.
There is one compatibility caveat in this change. We previously served
the graph view at the root path, without any redirects. If that old URL
is used with the new UI, it will now be the flame graph, changing the
meaning of the old URL.

I don't think there is a way around it and this does not seem to be an
outright bug, so I'm accepting it. But the new code makes it so that the
root URL redirects the browser to /flamegraph. This way any future
changes of the meaning of the root URL won't affect existing old URLs
since those will already have /flamegraph in them.

Fixes #950.